DjangoBackend/requirements.txt
Django==4.1.7
Pillow==9.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.14.0
